refactor(stats): move test stats to dedicated stats module
- Move get_test_stats() from services/test.py to services/stats/test.py
- Update stats module to export get_test_stats and get_test_run_stats
- Update router imports to use stats module for consistency
- Remove unused StatsCalculator import from services/test.py

This improves code organization by consolidating all statistics
functions (test, test_result, test_run) in the stats module, following
the established pattern and improving maintainability.
