+++ apps/backend/src/rhesis/backend/app/routers/test.py
@@ -10,7 +10,8 @@ from rhesis.backend.app.auth.user_utils import require_current_user_or_token
 from rhesis.backend.app.database import get_db
 from rhesis.backend.app.dependencies import get_tenant_context, get_db_session, get_tenant_db_session
 from rhesis.backend.app.models.user import User
-from rhesis.backend.app.services.test import bulk_create_tests, get_test_stats
+from rhesis.backend.app.services.stats import get_test_stats
+from rhesis.backend.app.services.test import bulk_create_tests
 from rhesis.backend.app.utils.database_exceptions import handle_database_exceptions
 from rhesis.backend.app.utils.decorators import with_count_header
 from rhesis.backend.app.utils.schema_factory import create_detailed_schema
