+++ apps/backend/src/rhesis/backend/app/services/stats/__init__.py
@@ -5,7 +5,9 @@ It includes support for entity statistics, test result analytics, and historical
 
 Main Components:
 - StatsCalculator: Core class for general entity statistics
+- get_test_stats: Specialized function for test entity statistics
 - get_test_result_stats: Specialized function for test result analytics
+- get_test_run_stats: Specialized function for test run analytics
 """
 
 # Core classes and configurations
@@ -13,7 +15,9 @@ from .calculator import StatsCalculator
 from .config import DimensionInfo, StatsConfig, StatsResult
 
 # Specialized functions
+from .test import get_test_stats
 from .test_result import get_test_result_stats
+from .test_run import get_test_run_stats
 from .utils import timer
 
 __all__ = [
@@ -25,5 +29,7 @@ __all__ = [
     # Utilities
     "timer",
     # Main functions
+    "get_test_stats",
     "get_test_result_stats",
+    "get_test_run_stats",
 ]
